+++ R/catalogs.R
@@ -14,7 +14,7 @@
 #' @source [Time-series bulk data download](https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html)
 #'
 #' @param catalog A single value indicating the catalogs to be updated
-#'   or "ALL" as a shorthand. See Details
+#'   or `"ALL"` as a shorthand. See **Details**.
 #'
 #' @param parse_dates Logical. If `TRUE` the dates would be parsed using
 #'  [bde_parse_dates()].
@@ -47,7 +47,7 @@
 #'
 #' ```
 #'
-#' Use "ALL" as a shorthand for updating all the catalogs at a glance.
+#' Use `"ALL"` as a shorthand for updating all the catalogs at a glance.
 #'
 #' If the requested catalog is not cached [bde_catalog_update()] is invoked.
 #'
@@ -206,10 +206,12 @@ bde_catalog_load <-
 #' @source [Time-series bulk data download](https://www.bde.es/webbde/en/estadis/infoest/descarga_series_temporales.html)
 #'
 #' @param catalog A vector of characters indicating the catalogs to be updated
-#'   or "ALL" as a shorthand. See Details
+#'   or `"ALL"` as a shorthand. See **Details**.
+#'
 #' @param cache_dir A path to a cache directory. The directory can also be set
 #'   via options with `options(bde_cache_dir = "path/to/dir")`.
-#' @param verbose Logical, display information useful for debugging.
+#' @param verbose Logical `TRUE` or `FALSE`, display information useful for
+#'   debugging.
 #'
 #' @details
 #' Accepted values for `catalog` are:
@@ -233,7 +235,7 @@ bde_catalog_load <-
 #'
 #' ```
 #'
-#' Use "ALL" as a shorthand for updating all the catalogs at a glance.
+#' Use `"ALL"` as a shorthand for updating all the catalogs at a glance.
 #'
 #' @examples
 #' \donttest{
@@ -314,7 +316,8 @@ bde_catalog_update <-
 #'
 #' @return A tibble with the results of the query.
 #'
-#' @param pattern regex pattern to search See Details and Examples.
+#' @param pattern [`regex`][base::grep()]  pattern to search See
+#'   **Details** and **Examples**.
 #'
 #' @inheritDotParams bde_catalog_load
 #'
@@ -325,11 +328,11 @@ bde_catalog_update <-
 #' Spanish, for the time being. Therefore search terms should be provided
 #' in Spanish as well in order to get search results.
 #'
-#' This function uses [base::grep()] function for finding matches on
-#' the catalogs. You can pass [regular expressions][base::grep()] to broaden
+#' This function uses [base::regex()] function for finding matches on
+#' the catalogs. You can pass [regular expressions][base::regex()] to broaden
 #' the search.
 #'
-#' @seealso [bde_catalog_load()], [base::grep()]
+#' @seealso [bde_catalog_load()], [base::regex()]
 #'
 #' @examples
 #' \donttest{