+++ R/series.R
@@ -9,9 +9,9 @@
 #'
 #' @encoding UTF-8
 #'
-#' @param series_code a numeric value or vector with time-series code(s), as
-#'   defined in the field `Número secuencial` of the corresponding
-#'   series. See [bde_catalog_load()].
+#' @param series_code a numeric (or coercible with [base::as.double()] value
+#'   or vector with time-series code(s), as defined in the field
+#'   `Número secuencial` of the corresponding series. See [bde_catalog_load()].
 #'
 #' @param series_label Optional. Character vector or value. Allows to specify a
 #'   custom label for the series extracted. It should have the same length than
@@ -74,11 +74,13 @@ bde_series_load <- function(series_code,
                             update_cache = FALSE,
                             verbose = FALSE,
                             extract_metadata = FALSE) {
-  if (missing(series_code) || !is.numeric(series_code)) {
-    stop("`series_code` must be numeric.")
+  if (missing(series_code)) {
+    stop("`series_code` can't be NULL")
   }
 
-  series_code <- unique(series_code)
+  series_code <- as.double(series_code)
+  # Remove NAs
+  series_code <- series_code[!is.na(series_code)]
 
   if (is.null(series_label)) {
     series_label <- as.character(series_code)
@@ -165,10 +167,6 @@ bde_series_load <- function(series_code,
 
 #' Load BdE full time-series files
 #'
-#' @export
-#'
-#'
-#' @description
 #' Load a full time-series file provided by BdE.
 #'
 #' ## About BdE file naming
@@ -180,6 +178,9 @@ bde_series_load <- function(series_code,
 #'
 #' For that reason, the function [bde_series_load()] is more suitable for
 #' extracting specific time-series.
+#'
+#'
+#' @export
 #'
 #' @concept series
 #'
@@ -194,7 +195,7 @@ bde_series_load <- function(series_code,
 #' @inheritParams bde_catalog_load
 #'
 #' @param parse_numeric Logical. If `TRUE` the columns would be parsed to
-#'   double (numeric) values. See Note.
+#'   double (numeric) values. See **Note**.
 #'
 #' @param extract_metadata Logical `TRUE/FALSE`. On `TRUE` the output is the
 #'   metadata of the requested series.