+++ NEWS.md
@@ -1,3 +1,9 @@
+# tidyBdE (development version)
+
+-   Update docs and vignettes.
+-   `bde_series_load()` now tries to coerce characters to numeric.
+
+
 # tidyBdE 0.2.0
 -   Add series index article.
 -   Add new series: `bde_ind_gdp_quarterly()`, `bde_ind_population()`.